requirements.txt
torch
torchvision
torchaudio
transformers
numpy 
scipy
sentencepiece==0.1.99
fastapi
uvicorn
openai
text2vec
gradio
pydantic
APScheduler
jieba
gensim
onnxruntime-gpu==1.15.1
pymilvus